Revert "remove redundant casts"
This reverts commit e7bf60e1d2f64b42431f9757e7be148f1d64af3d.

+++ src/main/java/com/projecturanus/betterp2p/util/GlStateManager.java
@@ -608,7 +608,7 @@ public class GlStateManager {
     }
 
     public static void glTexCoordPointer(int size, int type, int stride, int buffer_offset) {
-        GL11.glTexCoordPointer(size, type, stride, buffer_offset);
+        GL11.glTexCoordPointer(size, type, stride, (long) buffer_offset);
     }
 
     public static void glTexCoordPointer(int size, int type, int stride, ByteBuffer buffer) {
@@ -616,7 +616,7 @@ public class GlStateManager {
     }
 
     public static void glVertexPointer(int size, int type, int stride, int buffer_offset) {
-        GL11.glVertexPointer(size, type, stride, buffer_offset);
+        GL11.glVertexPointer(size, type, stride, (long) buffer_offset);
     }
 
     public static void glVertexPointer(int size, int type, int stride, ByteBuffer buffer) {
@@ -624,7 +624,7 @@ public class GlStateManager {
     }
 
     public static void glColorPointer(int size, int type, int stride, int buffer_offset) {
-        GL11.glColorPointer(size, type, stride, buffer_offset);
+        GL11.glColorPointer(size, type, stride, (long) buffer_offset);
     }
 
     public static void glColorPointer(int size, int type, int stride, ByteBuffer buffer) {
